Move schema types into their own modules
