+++ spatialos-sdk/src/worker/schema.rs
@@ -8,204 +8,16 @@
 //! * There are two Rust types for each C type: An immutable, borrowed version; And
 //!   a mutable, owned version.
 
-use crate::worker::{
-    component::{Component, ComponentId},
-    internal::Sealed,
-    EntityId,
-};
+use crate::worker::{internal::Sealed, EntityId};
 use spatialos_sdk_sys::worker::*;
-use std::{
-    collections::BTreeMap,
-    mem,
-    ops::{Deref, DerefMut},
-    ptr::NonNull,
-    slice,
-};
+use std::{collections::BTreeMap, mem, slice};
+
+mod component_data;
+mod object;
+
+pub use self::{component_data::*, object::*};
 
 pub type FieldId = u32;
-
-/// Serialized schema data for a component, owned by the Rust SDK.
-///
-/// For maximum efficiency, the serialized data may borrow data from the component
-/// used to create an `OwnedComponentData` instance. The lifetime parameter
-/// tracks this borrow, such that an `OwnedComponentData` cannot outlive the
-/// data it borrows.
-#[derive(Debug)]
-pub struct ComponentData {
-    raw: NonNull<Schema_ComponentData>,
-}
-
-impl ComponentData {
-    pub fn new<C: Component>(component: &C) -> Self {
-        // Create the underlying `Schema_ComponentData` and retrieve the fields object.
-        let raw = NonNull::new(unsafe { Schema_CreateComponentData(C::ID) }).unwrap();
-        let mut result = Self { raw };
-
-        // Populate the schema data from the component.
-        component.into_object(&mut result.fields_mut());
-
-        result
-    }
-
-    pub fn component_id(&self) -> ComponentId {
-        unsafe { Schema_GetComponentDataComponentId(self.raw.as_ref()) }
-    }
-
-    pub fn fields(&self) -> ObjectRef<'_> {
-        self.as_ref().fields()
-    }
-
-    pub fn fields_mut(&mut self) -> ObjectMut<'_> {
-        unsafe { ObjectMut::from_raw(&mut *Schema_GetComponentDataFields(self.raw.as_ptr())) }
-    }
-
-    pub fn as_ref(&self) -> ComponentDataRef<'_> {
-        ComponentDataRef {
-            raw: unsafe { self.raw.as_ref() },
-        }
-    }
-
-    pub fn deserialize<T: SchemaObjectType>(&self) -> T {
-        T::from_object(self.fields())
-    }
-
-    /// Converts the `OwnedComponentData` into a `*mut Schema_ComponentData` that can be
-    /// passed to the C API.
-    ///
-    /// This transfers ownership of the data to the caller, so the caller needs to
-    /// ensure that the appropriate steps are taken to free the data. If the raw data is
-    /// passed to the C API, the C SDK will take ownership of the data and will free it
-    /// when it's done.
-    pub fn into_raw(self) -> *mut Schema_ComponentData {
-        let raw = self.raw;
-        mem::forget(self);
-        raw.as_ptr()
-    }
-}
-
-impl Drop for ComponentData {
-    fn drop(&mut self) {
-        unsafe {
-            Schema_DestroyComponentData(self.raw.as_ptr());
-        }
-    }
-}
-
-/// Serialized schema data for a compnent owned by the C SDK.
-///
-/// The lifetime parameter tracks the parent data that owns the schema data
-/// (generally an `OpList`), such the `SchemaComponentData` instance cannot live
-/// its parent.
-#[derive(Debug)]
-pub struct ComponentDataRef<'owner> {
-    raw: &'owner Schema_ComponentData,
-}
-
-impl<'owner> ComponentDataRef<'owner> {
-    pub unsafe fn from_raw(raw: &'owner Schema_ComponentData) -> Self {
-        Self { raw }
-    }
-
-    pub fn component_id(&self) -> ComponentId {
-        unsafe { Schema_GetComponentDataComponentId(self.raw) }
-    }
-
-    pub fn fields(&self) -> ObjectRef<'owner> {
-        unsafe {
-            ObjectRef::from_raw(&*Schema_GetComponentDataFields(
-                self.raw as *const _ as *mut _,
-            ))
-        }
-    }
-
-    pub fn deserialize<T: SchemaObjectType>(&self) -> T {
-        T::from_object(self.fields())
-    }
-}
-
-#[derive(Debug)]
-pub struct ObjectMut<'owner> {
-    raw: &'owner mut Schema_Object,
-}
-
-impl<'owner> ObjectMut<'owner> {
-    pub(crate) unsafe fn from_raw(raw: &'owner mut Schema_Object) -> Self {
-        Self { raw }
-    }
-
-    pub fn add_field<T: SchemaField>(&mut self, field: FieldId, value: &T::RustType) {
-        T::add_field(self, field, value);
-    }
-
-    pub fn add_field_array<T: ArrayField>(&mut self, field: FieldId, value: &[T::RustType]) {
-        T::add_field_list(self, field, value);
-    }
-
-    pub fn as_ref(&'owner self) -> ObjectRef<'owner> {
-        ObjectRef { raw: self.raw }
-    }
-}
-
-impl<'a> AsRef<Schema_Object> for ObjectMut<'a> {
-    fn as_ref(&self) -> &Schema_Object {
-        self.raw
-    }
-}
-
-impl<'a> AsMut<Schema_Object> for ObjectMut<'a> {
-    fn as_mut(&mut self) -> &mut Schema_Object {
-        self.raw
-    }
-}
-
-impl<'a> Deref for ObjectMut<'a> {
-    type Target = Schema_Object;
-
-    fn deref(&self) -> &Self::Target {
-        self.raw
-    }
-}
-
-impl<'a> DerefMut for ObjectMut<'a> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.raw
-    }
-}
-
-#[derive(Debug, Clone, Copy)]
-pub struct ObjectRef<'a> {
-    raw: &'a Schema_Object,
-}
-
-impl<'a> ObjectRef<'a> {
-    pub unsafe fn from_raw(raw: &'a Schema_Object) -> Self {
-        Self { raw }
-    }
-
-    pub fn field<T: SchemaField>(self, field: FieldId) -> T::RustType {
-        T::get_field(self, field)
-    }
-
-    pub fn field_array<T: ArrayField>(self, field: FieldId) -> Vec<T::RustType> {
-        let mut result = Vec::new();
-        T::get_field_list(self, field, &mut result);
-        result
-    }
-}
-
-impl<'a> AsRef<Schema_Object> for ObjectRef<'a> {
-    fn as_ref(&self) -> &Schema_Object {
-        self.raw
-    }
-}
-
-impl<'a> Deref for ObjectRef<'a> {
-    type Target = Schema_Object;
-
-    fn deref(&self) -> &Self::Target {
-        self.raw
-    }
-}
 
 // =================================================================================================
 // Schema Conversion Traits
@@ -244,27 +56,26 @@ impl<T: SchemaObjectType> SchemaField for T {
 
     fn add_field(object: &mut ObjectMut<'_>, field: FieldId, value: &Self::RustType) {
         let mut field_object =
-            unsafe { ObjectMut::from_raw(&mut *Schema_AddObject(object.raw, field)) };
+            unsafe { ObjectMut::from_raw(&mut *Schema_AddObject(object.as_mut(), field)) };
         value.into_object(&mut field_object);
     }
 
     fn get_field(object: ObjectRef<'_>, field: FieldId) -> Self::RustType {
-        let field_object = unsafe {
-            ObjectRef::from_raw(&*Schema_GetObject(object.raw as *const _ as *mut _, field))
-        };
+        let field_object =
+            unsafe { ObjectRef::from_raw(&*Schema_GetObject(object.as_ptr() as *mut _, field)) };
         T::from_object(field_object)
     }
 }
 
 impl<T: SchemaObjectType> IndexedField for T {
     fn field_count(object: ObjectRef<'_>, field: FieldId) -> u32 {
-        unsafe { Schema_GetObjectCount(object.raw, field) }
+        unsafe { Schema_GetObjectCount(object.as_ptr(), field) }
     }
 
     fn index_field(object: ObjectRef<'_>, field: FieldId, index: u32) -> Self::RustType {
         let field_object = unsafe {
             ObjectRef::from_raw(&*Schema_IndexObject(
-                object.raw as *const _ as *mut _,
+                object.as_ptr() as *const _ as *mut _,
                 field,
                 index,
             ))
@@ -340,24 +151,24 @@ macro_rules! impl_primitive_field {
             ) {
                 let value = (*value).into();
                 unsafe {
-                    $schema_add(object.raw, field, value);
+                    $schema_add(object.as_ptr(), field, value);
                 }
             }
 
             #[allow(clippy::useless_transmute, clippy::transmute_int_to_bool)]
             fn get_field(object: ObjectRef, field: FieldId) -> Self::RustType {
-                unsafe { mem::transmute($schema_get(object.raw, field)) }
+                unsafe { mem::transmute($schema_get(object.as_ptr(), field)) }
             }
         }
 
         impl IndexedField for $schema_type {
             fn field_count(object: ObjectRef<'_>, field: FieldId) -> u32 {
-                unsafe { $schema_count(object.raw, field) }
+                unsafe { $schema_count(object.as_ptr(), field) }
             }
 
             #[allow(clippy::useless_transmute, clippy::transmute_int_to_bool)]
             fn index_field(object: ObjectRef<'_>, field: FieldId, index: u32) -> Self::RustType {
-                unsafe { mem::transmute($schema_index(object.raw, field, index)) }
+                unsafe { mem::transmute($schema_index(object.as_ptr(), field, index)) }
             }
         }
 
@@ -389,7 +200,7 @@ macro_rules! impl_primitive_field {
 
                 unsafe {
                     $schema_add_list(
-                        object.raw,
+                        object.as_ptr(),
                         field,
                         buffer.as_ptr() as *const _,
                         buffer.len() as u32,
@@ -412,7 +223,7 @@ macro_rules! impl_primitive_field {
                 // Replace the contents of `data` with the list of values in the schema field.
                 unsafe {
                     data.set_len(count);
-                    $schema_get_list(object.raw, field, data.as_mut_ptr() as *mut _);
+                    $schema_get_list(object.as_ptr(), field, data.as_mut_ptr() as *mut _);
                 }
             }
         }
@@ -583,13 +394,13 @@ impl<T: IndexedField> Sealed for Option<T> {}
 impl<T: IndexedField> SchemaField for Option<T> {
     type RustType = Option<T::RustType>;
 
-    fn add_field<'owner>(object: &mut ObjectMut<'owner>, field: FieldId, value: &Self::RustType) {
+    fn add_field(object: &mut ObjectMut<'_>, field: FieldId, value: &Self::RustType) {
         if let Some(value) = value {
             T::add_field(object, field, value);
         }
     }
 
-    fn get_field<'owner>(object: ObjectRef<'owner>, field: FieldId) -> Self::RustType {
+    fn get_field(object: ObjectRef<'_>, field: FieldId) -> Self::RustType {
         match T::field_count(object, field) {
             0 => None,
             _ => Some(T::get_field(object, field)),
@@ -617,7 +428,7 @@ where
         // Create a key-value pair object for each entry in the map.
         for (key, value) in map {
             let mut pair =
-                unsafe { ObjectMut::from_raw(&mut *Schema_AddObject(object.raw, field)) };
+                unsafe { ObjectMut::from_raw(&mut *Schema_AddObject(object.as_ptr(), field)) };
             pair.add_field::<K>(SCHEMA_MAP_KEY_FIELD_ID, key);
             pair.add_field::<V>(SCHEMA_MAP_VALUE_FIELD_ID, value);
         }
@@ -630,7 +441,7 @@ where
         for index in 0..count {
             let pair = unsafe {
                 ObjectRef::from_raw(&*Schema_IndexObject(
-                    object.raw as *const _ as *mut _,
+                    object.as_ptr() as *const _ as *mut _,
                     field,
                     index,
                 ))
@@ -686,7 +497,7 @@ impl SchemaField for String {
 
 impl IndexedField for String {
     fn field_count(object: ObjectRef, field: FieldId) -> u32 {
-        unsafe { Schema_GetBytesCount(object.raw, field) }
+        unsafe { Schema_GetBytesCount(object.as_ptr(), field) }
     }
 
     fn index_field(object: ObjectRef, field: FieldId, index: u32) -> Self::RustType {
@@ -713,7 +524,7 @@ impl SchemaField for Vec<u8> {
 
 impl IndexedField for Vec<u8> {
     fn field_count(object: ObjectRef<'_>, field: FieldId) -> u32 {
-        unsafe { Schema_GetBytesCount(object.raw, field) }
+        unsafe { Schema_GetBytesCount(object.as_ptr(), field) }
     }
 
     fn index_field(object: ObjectRef<'_>, field: FieldId, index: u32) -> Self::RustType {
@@ -731,22 +542,22 @@ fn add_bytes<'owner>(object: &mut ObjectMut<'owner>, field: FieldId, bytes: &[u8
 
     // Add `buffer` to `object` as the field.
     unsafe {
-        Schema_AddBytes(object.raw, field, buffer.as_ptr(), buffer.len() as _);
+        Schema_AddBytes(object.as_ptr(), field, buffer.as_ptr(), buffer.len() as _);
     }
 }
 
 fn get_bytes(object: ObjectRef<'_>, field: FieldId) -> &[u8] {
     unsafe {
-        let data = Schema_GetBytes(object.raw, field);
-        let len = Schema_GetBytesLength(object.raw, field);
+        let data = Schema_GetBytes(object.as_ptr(), field);
+        let len = Schema_GetBytesLength(object.as_ptr(), field);
         std::slice::from_raw_parts(data, len as usize)
     }
 }
 
 fn index_bytes(object: ObjectRef<'_>, field: FieldId, index: u32) -> &[u8] {
     unsafe {
-        let data = Schema_IndexBytes(object.raw, field, index);
-        let len = Schema_IndexBytesLength(object.raw, field, index);
+        let data = Schema_IndexBytes(object.as_ptr(), field, index);
+        let len = Schema_IndexBytesLength(object.as_ptr(), field, index);
         std::slice::from_raw_parts(data, len as usize)
     }
 }
